Implement SSA split; restore speculative args opt
There were two ways to try and do this: rename everything looking
back, or rename everything going forward. Going forward seems to be
rather simpler to make work, especially since the tree going forward
has not yet been optimized. So, that's done. Also if we do miss any
renames, it's fairly harmless, because they would just miss the
facts. (Note that the facts are not being added, for the reasons
that triggered this work.)