Don't add base index when looking for lexical type
This information is looked up against the static frame, while it's
only the dynamic frame that is subject to the base_index due to the
potential for inlining. Fixes a read past the end of the static frame
lexical registry buffer.

Also move `lexreg` back outside of the loop; it's not managed by the
GC and so can't move.