up to double-devirt assign_[ins] and decont_[ins]
either directly invoke the functions from the container
spec, or - for native refs - go one step further and
decide which function to call based on the kind of native
ref, i.e. lexical, attribute, positional, or multidim.

The benchmark I started with now runs in 90% of its
original time; it was just decont_i, add 1, assign_i in
a tight loop.