Fix segfaults in GC caused by sharing of jitcode in NativeCall repr
jitcode is a pointer to an MVMJitCode structure. When copying a NativeCall
object, we may not just copy the pointer but have to make a copy of the whole
structure. Otherwise GC'd free the same structure multiple times.

Many thanks to dogbert17++ for providing the ASAN printout that gave the
essential hint!

Fixes GH #862