Fix a regression with Regional Indicator Unicode Graph Seg.
The state for when past regional indicators were encountered would last
too long, leading segmentation code breaking in the wrong place when
multiple Regional Indicators were encountered that were not contiguous.
(RI) (RI) (RI) would work fine and become (RI)(RI) (RI) (first two are one
grapheme while the third one is by itself).

If we had instead (RI) (RI) (h) (i) (RI) (RI) this would become
(RI)(RI) (h) (i) (RI) (RI) instead of (RI)(RI) (h) (i) (RI)(RI)
as it should be (here, if two parenthized terms are touching that means
they are part of the same grapheme).

I also fixed bugs when join was used with regional indicators as well as
the reported bug in MVM_string_concatenate().

This also makes the state a little simpler, either 0 or 1 in
norm->regional_indicator. And when we try and check if we can concat or
join two strings, we preset norm->regional_indicator=1 to pretend that the
unknown codepoint before the one to be joined is a regional indicator.
This way we can detect cases when we need full renormalization more
consistantly, while at the same time making sure we reset
norm->regional_indicator when we see a non-Regional Indicator codepoint.