Immediately promote objects ref'd from gen2
When we hit the write barrier, annotate on the object that was being
referenced that it was pointed to by a gen2 object. Then, if it is
alive at the next GC, immediately promote it to gen2 rather than it
having another stay in the nursery.

This makes a "populate a million element array" benchmark run in 90%
of the time that it used to take, and will help other programs that
build up long-lived data structures. CORE.setting compilation is one
such program, and seems to also benefit from this. Meanwhile, other
benchmarks seem to not be negatively impacted by this.