[Spesh] Destroy inlinee after inlining
After we've inlined a graph the inliner claims ownership over the memory
of the inlinee. With this patch, we destroy the original graph, which
should simplify lifetime management. The memory region list is merged with
the inliners graph so that the memory allocated for the inlinee nodes
is not freed.