Re-instate use of int cache in fast boxing
For JIT, we now emit assembly that does the comparisons and then does
a fetch from the integer cache if applicable. Thus the only real call
is to obtain memory from the GC if we need to box, the value then
being poked into the object at the appropriate location.

Also stub in as-yet unimplemented and unused spesh ops that we'll use
for optimizing boxing to a P6bigint embedded in a P6opaque.