Lower add_I, sub_I, and mul_I where possible
When the input and output types are consistent (which should be the
overwhelmingly common case) we JIT-compile these into code that tries
to do the operation directly if we are dealing with two smallint input
values, and provided it doesn't overflow stores it back. If either of
those two conditions isn't met, it falls back to a slow path. Since we
have access to the CPU instruction set, we can use its overflow flag
rather than having to bounds-check the result.

The allocation of the result is also done through fastcreate, which is
a further win. A simple loop doing 10 million additions goes down to
73% of its original runtime, while a benchmark doing reading of lines
and adding up the number of chars in an Int runs in 93% of the original
runtime.