Correct tracking of log guards at PHI nodes
Previously, we only had them on the facts of the PHI node itself.
However, those facts could then be copied and relied on elsewhere. With
this change, we now keep a list of the log guards that we depend on for
each fact, and at a PHI node we copy all the input lists into a target
list. We then, when we depend on or copy facts, just assign the input
log guards list and count (which is safe since it's immutable, and we
use the region allocator so there's no memory management concerns).