Fix rw args of JITed calls to JITed native calls
getarg_* are used to read (possibly modified) values back from the args buffer.
This is currently done only for native calls (after a nativeinvoke). However
the JIT does not emit any code for the corresponding arg_* ops for a
nativeinvoke. Instead the values are read directly from the WORK registers and
the arg_* ops are only used to mark those registers. This means that there's
nothing in the args buffer to read from.

So instead, turn the getarg_* ops into plain set ops and take the value from
the original WORK register.