Fix unshift_* templates (no destructive '!')
Remove destructive '!' character. As bdw explained in his feedback, the
'!' identifies an op that doesn't only fail to produce a value, but also
has a write register that is explicitly populated within the template.

Since the unshift ops don't write to a register, it is not correct to
mark it with '!'.