[UNI] Make Block name lookup work for unassigned cp's
Unlike other properties which if they are unassigned go to default
values, Block does not work this way. Instead of looking it up in the
database, use similar code to how we check the block in other
circumstances and use binary search over our Block info data structure.

This fixes issue #924

TODO: Remove the block data from the database. This requires a fair
amount of effort because all of our data is linked to having items in
the database, so for now fix it without removing the now unused data
from the database.