[ConcBlockingQueue] Implement unshift
Enable queue users to 'jump the line' and make sure that their item is
delivered to the front of the queue. This has less concurrency than
the push/shift pair, because it needs to lock both the head and tail
locks. It should rarely be useful.

The motivating example in which it is, is shutting down the spesh
worker thread, because it means we do not have to wait until the spesh
thread has processed (now useless) work.