Fix alignment of nursery allocations
Recent changes used a path to nursery allocation that are not aligned,
leading to segfaults on e.g. armhf. Alignment is really a property of
the nursery as it is also encoded in the collector, so doing the
alignment within the allocator, not at the call site.

Fixes https://github.com/MoarVM/MoarVM/issues/951