Fix wrong field in exprjit cast and add macro
In the 'objprimbits' template, accidently used the wrong field name for
cast's from_size value ('boxed_primitive' instead of 'bits').

In this case, 'boxed_primitive' and 'bits' are both the same size, but
that doesn't make it less wrong just less painfully wrong 😉.

Added a macro that makes this idiom more DRY and less prone to this type
of error.