[IO] Manage event loop uv_loop_t explicitly
Because we want to be able to start and destroy the IO loop thread, we
cannot use the threadcontexts' uv_loop_t structure, but should use a
per-instance structure instead.

Also, we can initialize the async handler directly - no need to wait
for the thread to start. (Next up - don't wait for the thread to
start).

MVMAsyncOps handlers should use the provided uv_loop_t; found and
fixed a number of instances where it wasn't.