Add support for utf-16 decodestream
Adds support for utf-16. At the moment only handles the same endianess
as the processor. We may want to have utf16le and utf16be as options.

When we decode a utf16 string (standard, not decodestream) we check for
a BOM, byte order marker. But if we are using decodestream we may be
starting decode partway through the file, so we will probably have to
save this data.

For now this gets utf16 working for those who need it.