Add utf16le and utf16be encoding types and impl. for buf decode
We add utf16 little endian and utf16 big endian as encoding types.
For now it is only implemented for buffer decode.

At the moment these new encodings will ignore and pass through any
byte order marker (BOM) as stated in RFC 2781.
Most programs will delete the BOM even when you set it to utf16be or
utf16le, so we may decide to delete the BOM if it exists at the
beginning of the buffer. This functionality is not at all standardized
for MoarVM yet.

In addition to the encoding additions, I do some refactoring of
initializing the encoding name's MVMString's so that we can more easily
add new encodings by adding it to the new struct; thus getting rid of
some big case statements and if else chains in initialization and
conversion from MVMString name to the internal encoding number.