Fix some bugs with utf16* decodestream get_chars, and retain state
Retain state i.e. which encoding we are currently using. This is
important for utf16 encoding since it detects using a byte order mark
whether it is big or little endian. We must save this detection so
further calls use the proper encoding.

Also fix issues with get_chars, which the location could get off if you
were reading one unicode codepoint at a time instead of slurping
everything.