Don't set ref'd from gen2 flag during GC
Parallel GC relies on us only updating the object flags from the thread
that owns the object (or the nominated thread to GC that thread's
objects). The updates caused by the write barrier touching the object
that was referenced caused races that could lead to incorrect GC
behavior, which typically showed up as an invalid SC (which was really
because the forwarder lives in the same slot that the SC used to on a
moved object).