Pass through BOM with utf16le and utf16be
I have decided for now to act similar to Perl 5 and for writing and
reading to utf16le and utf16be it will not act any differently if there
is a BOM encountered.

So it will pass through the BOM and zero width non-breaking space for
these encodings. For utf16 it will follow and remove the BOM as it did
before.

Still to be solved: utf16 has not outputted a BOM previously, and I will
need to decide if adding a BOM will cause problems for backward
compatibility.