Perform parameter logging at entry time
Rather than doing it for each of the parameter instructions. This lets
us log parameter types even when their target will be a slurpy, meaning
we can make a type specialization rather than a certain specialization
for calls to things taking slurpy parameters. This helps, among likely
quite a few things, calls like `.new(x => 1, y => 2)` for the standard
`new` constructor: when the `BUILDALL` is small enough, it can now be
inlined into the specialization of `new`.