Simplify/improve optimization of decont
We don't need to track "is this decontainerized" everywhere; we can
just look at the facts about the type to see if it's a container type
or not. By doing so, we spot optimize some cases where we forgot to
set the "it's decontainerized" flag, so this is not only a cleanup,
but also a performance win.