Fix poor inlining due to wrong plan sorting
We sorted the plan by stack depth, which is a decent approximation,
but sometimes we could have a situation where a caller and callee
planned for specialization would end up with the caller having a
greater maximum observed stack depth than the callee.

If we detect such a case, give the callee one a boost, so that we'll
do them in a more sensible order, leading to more inlining chances.