Update Unicode to 11.0 and update grapheme break rules
Update to the 11.0 version of the Unicode database. Get rid of multiple
old rules using now obsolete Emoji rules/guidelines for favor of the new
Extended_Pictographic property. We don't pass three of the Unicode
grapheme break tests 11.0, though this is about where we were for the
Unicode 10.0 tests and is acceptable enough to change versions.