Introduce irreplaceable flag for allocations
Which indicates that we can't currently scalar replace them. In the
future, we will often be able to just materialize the object at that
point in time, if it makes sense to do so. For now, this means we can
start to liberalize things a little bit. The most immediate way is to
not bail the whole analysis just because one tracked allocation cannot
be scalar replaced.

Additionally, properly handle PHIs that sneak in mid-block (due to
other opts) and the argument to an object attribute bind.