[JIT] Do not skip adding labels to PHI nodes
In some circumstances, apparently we can end up with PHI nodes with
labels on them that are not at the start of the tree. These labels
would be allocated but not emitted, which would (in this case) lead to
incorrect exception handler bounds.

DynASM reported this but this (exceptional and very wrong) condition
was previously being logged only to the JIT log, where nobody ever saw
it. When I changed it to stderr, reports came in, and now it's fixed.