Handle non-merge case of PHI
Sometimes we end up with a PHI that only actually has one soure to
merge. In this case, we can track allocations through it.

These show up when a call is inlined, since the return value is always
put through a PHI (as part of handling multiple returns consistently).
Thus this means we can now handle some simple cases of eliminating an
allocation that is made in one routine and immediately got rid of in
another.