introducing vectorapply for native arrays
Just a function with a bunch of very tight loops
that the C compiler will turn into vectorized code.

Has the same semantics as the Z and X operators. Pass
in three native arrays of the same slot type, the
moarvm opcode you want to have vectorized (currently
only (add|mul|sub)_[ni] are implemented) and whether
to use "cross" semantics (only crossing with a single
value is currently implemented).