Seed the deserialization string heap for getting correct indexes
Originally there were two pieces of code for collecting the string heap:
the data serialization (nqp::serialize) and the MAST compiler. The strings
collected during serialization were used as seed for the MAST compiler's
string heap. MVM_serialization_serialize assumed that it was passed an
empty string heap to fill.
Since we now assemble bytecode (and the string heap) before serializing
data, we have to flip this. Serialization's string heap is now seeded with
the bytecode compiler's.