[JIT] Fix trunc to cast upwards
This is a bit confusing altogether.

My original theory of trunc was that we'd do a downcast, clearing the
upper bits, storing only the lower bits, but treating the number as
being smaller. But all sorts of downstream code didn't really agree.

So now we do the opposite - zero-extend a smaller number into a bigger
one (and store that instead). And that seems to fix a few bugs.