Always rebuild minilua to work around issue #999
This ensures that minilua will always be rebuild. This fixes bug #999,
though we may find a better solution.

We declare ALWAYS_REBUILD as a target, and use the .PHONY property to
designate that it's not an actual file (though it's unlikely that file
would exist on the users computer). While we're there, also add all the
other phony rules like 'install' 'clean' 'lib' etc. as .PHONY so the
rules won't break if those files happen to exist in that folder.