free spesh log entries after consuming them
this prevents threads that do little to no
allocations but lots of spesh logging in an
otherwise idle program from growing to multiple
gigabytes of ram used solely for spesh log entries.

This happened with rakudo's ThreadPoolScheduler
supervisor thread when a program with multithreading
or async workloads was left idle, like for example
a Cro app.