Make space for BINARY_ENDIAN_(LITTLE|BIG|NATIVE) in (write|read)(u?int|num) ops
I missed the part of the specification that besides using little and big endian
allowed for just using native encoding. This means a single bit is not
sufficient and we need two bits for encoding endianness. Therefore we need to
adjust the size values.
To get around bootstrap issues we treat the size value different depending on
the little endian flag being set, as all current users are missing that but are
going to set it.