Fix repossessions that cause object size "changes"
This is only an issue for P6opaque, where a change of type might cause
a change of size. Thus, we need to allocate and use a new region for
the object attributes to be held in. Prior to this change we failed to
do that, which meant the deserialization would write over memory that
followed the original, now under-size, object body. This was often the
header of another object, and so tended to blow up much later when the
GC was touching the object, rather than anywhere close to the point
things actually went wrong.

This addresses https://github.com/rakudo/rakudo/issues/2519.