Remove wraparound handling of negative offsets in (read|write)_buf
Treating negative offsets as counting from the end of the array was meant to
provide consistency with other array ops. Turns out however that the right
course of action may be to remove this magic from other ops, too. The magic
does not reflect the semantics, Perl 6 actually needs.
Fixes GH #1024