Rename optimize_coerce To _in, Add coerce_ni
in cases where a coerce_ni is used on the result of
a coerce_in, we can just use the original integer.

Normally, the combination of these ops would cause
truncation of integers to only ~52 bits of precision,
but for many operations, like elems, it's likely that
the full integer was meant to be used.

This optimization may want to be limited to nqp, or
perhaps limited to the result of certain ops.