Teach optimize_smart_coerce About Boot Types
BOOTInt, BOOTNum, and BOOTStr wouldn't be
recognized to not have a Str or Num method, but
since we made them, we can be sure (if the user
didn't meddle).

This lets us optimize a bunch of smrt_numify and
smrt_strify into unboxing or explicit coercion
ops, which are faster than the more general smart
versions of coercing

Probably only ever useful in nqp code, I don't think
rakudo ever outputs this op?