Fix JIT compiled getcodename not always dieing on type objects
In cases where we do get a code object, we never checked if the code object
is concrete because we immediately jumped to the success branch. Invert the
logic and jump directly to throwing the exception instead to give the other
check an opportunity to fail.

Fixes "no Segfault when creating a Sub object with .bless and trying to say
it" in t/spec/S06-other/misc.t failing due to no exception being thrown.