Optimize band_i And bor_i With if_i And unless_i
a band_i or bor_i with two known inputs will be
constant-folded, and an if_i or unless_i acting
on a bor_i or band_i with one known input will
either be constant-folded (x and 0 always equals 0,
x or 1 always equals 1) or use the unknown input
parameter (x and 1 always equals x, x or 0 always
equals x).

The latter optimization can only be done when
the writer of the values is known to only ever
produce a 0 or a 1.