Save deopt usage info when writing bytecode
When we encounter an instruction and the register it writes has depot
users, then persist the bytecode location and the using deopt indices.
This is so we'll be able to reconstruct it when doing inlining.

A future improvement would be to only keep it when the bytecode is
within the inline size limit, using the same rules as the inliner itself
uses to make that decision.