Allow dead instruction elimination in inlines
Previously we could not do this, since we lacked any sense of which
values have to be preserved in order to not break deopt. Now that info
is available, we can perform this inside of inlines too. This is most
immediately useful when we consider instructions that produce a return
value, which is the inlined into something that doesn't care for the
return value. Such instructions can then simply be deleted.