Add a limit on spesh plugin guard set size
Above a certain size, it's both likely to be a spesh plugin bug, but
also likely to not be an optimization at all, so better to just run
the plugin each time. Add a mode (off by default) where we dump the
stack trace of the location where we were trying to add way too many
guards to the set, to help with tracking down such bugs.