Use correct static frame in plugin resolution
When we inline a speshresolve that needs late-bound lookups, we include
the static frame, meaning that we can then inline it. However, part of
the spesh plugin handling code just read the static frame of the frame
currently on the callstack, rather than using the one it was provided
with. If the plugin got new guards, they were resolved in the provided
static frame, but the new ones would be stored in the incorrect
(currently executing) static frame. This would lead to repeated addition
of the same guard to the wrong place, since it would be missed every
time.