Enable optimization of speshresolve with ustable guard indexes
Even if there's no clearly winning guard index, we can at least replace
speshresolve by sp_speshresolve and get rid of the logging and enable JIT
compilation for that frame.
The comments suggest that it has always been intended this way and that the
return was just an oversight happening due to the 2 halfs of the function
being written a month apart.