Allow CStruct pass-by-value in Nativecall
This feature can be implemented only on libffi.

This is the initial implementation of the pass-by-value
functionality for nativecall, it recognizes the 'copy' key on
parameter/return hashes. Then it build up the libffi parameter
definition by analyzing the STables and REPRs of the members. It
currently works only on P6num P6int and CStruct member types but
will be extended upon review. This functionality can be implemented
only on the libffi nativecall backend, and might render dyncall
obsolete for MoarVM.