Call strerror for directory errors
So that we get an error text rather than an error code, as reported in
Rakudo issue https://github.com/rakudo/rakudo/issues/2649. Note that
this only fixes it for the POSIX case; the Windows case probably also
needs a look into, to see how it should be done there.