Add deopt uses as needed for scalar replacements
When we perform scalar replacement, then we place what used to be in
object attributes into registers. If we deopt, we materialize the object
from those registers. However, if those register's only use is for
deopt, then we could delete writes into them, and thus would not be able
to do that materialization. Thus, be sure to add deopt usages to the
scalar replacement registers as required (we do it per version, so this
is quite precise). That way, they will be preserved as needed, and not
break our ability to deoptimize.