Copy deopt usage info when doing guard reg split
When we insert a guard, we do an SSA split of the pre-guard and
post-guard versions. We previously moved the list of deopt indexes that
require the register to the preguard register. However, they are also
needed on the postguard register too, since that needs to live over any
following deopts. Moving them caused the escape analyzer to fail to add
required entries into the deopt/materializations table, which in turn
led to broken state after deopt in some cases. Copying them satisfies
all situations.