Avoid duplicate checks in hash access
We would check that a string was of the correct representation and
concrete up to 3 times during the process of doing at_key or bind_key.
Furthermore, we compared the hash value a second time when we got a
potential hash match, since that was done in MVM_string_equal. Get rid
of this duplicate logic, which it seems a C compiler can't reasonably
eliminate. In a Perl 6 hash benchmark, this gave around 5% less CPU
cycles according to measurements taken with callgrind.