Add a newmixintype op
A form of `newtype` that sets a flag saying that this type can be the
target of a mixin. Added as a new op used at type creation time so that
we can easily reason about a type always having that property for its
lifetime, rather than any timing worries about it being set later. This
alone doesn't do anything useful; the useful part will be when we start
to enforce its usage and then introduce new ops that avoid the whole
"look at real_data" dance in P6opaque attribute access, which it seems
can save us around 4 CPU instructions per attribute read/write. Which,
given we do a lot of those - Perl 6 is full of objects - is potentially
going to be a measurable improvement.