Fix rw args of native calls not getting updated if JIT compilation fails
This fixes issue with rw args to native calls, when the JIT fails to compile
the call graph the nativeinvoke_o is in. Happens for example because we're
missing the case for MVM_OP_getlexref_u8.

optimize_getarg worked on the assumption that the JIT compiler will not emit any
code for the corresponding arg_* ops for a nativeinvoke and instead the values
would be taken from the WORK register. But in our case, the JIT bailed and the
arg_i was still there. So the address of the arg buffer was passed to the native
function and this is where the set value ended up in. But optimize_getarg had
turned the getarg_i op into a set from the WORK register, ignoring the value in
the arg buffer.

In other words: we handled the interpreted version and the speshed + JIT
compiled version but not the speshed version without JIT compilation.

As spesh cannot know in advance whether JIT compilation will succeed, move the
getarg_* optimization into the JIT compiler itself. This way we only reach it
if the previous ops could be compiled successfully. Now we're back to just 2
cases: either JIT compilation is successful and arg_* and getarg_* are both
removed (with the native code working directly with the WORK register), or both
are still there and interpreted (with the native code working on the ARGS
buffer)

Hopefully fixes GH #2706