Fix nativecall refresh clearing used child object pointers
We compared the address of the struct member pointing to a child object with
the address of the actual child object. Of course this cannot match, so we
ended up clearing the child_objs slot all the time. This caused us to garbage
collect e.g. a CArray object that was still pointed to by the CStruct.