Improve check for used toolchain on macOS
Don't strictly test for make compliance because XCode is using GNU make
though earlier version of it. This shouldn't break the build process
unless a feature from later version of GNU make would make it into the
Makefile which won't work in XCode's version of make. But this would
draw the whole XCode build broken anyway.

Still, to be on the safe side, this commit makes Configure.pl forcingly
choose /usr/bin/make (which always belongs to XCode) unless GNU
toolchain is requested.