Avoid preserving ->caller unless we really need it
We need it for backtraces and context introspection. Otherwise, it can
go. Preserving it can cause leaks involving taken closures: the taking
of the closure has an ->outer frame that in turn points to a ->caller
and keeps it alive. This was involved in the leak that was reported in
https://github.com/rakudo/rakudo/issues/2803.