Handle inlining of outer lexical binds
This is used in code using native types, and was one source of surprise
slowness when using natives. For example, this:

    my num $x = 1.5e0;
    my num $y = 0e0;
    for ^10_000_000 {
        $y = $y + $x;
    }

Used to run slower than if the native types were not added. Now it runs
in 25% of the time it did before the change, meaning it pulls well ahead
of the case without native types (although EA will, in the future,
likely close down that gap again).