First steps towards SR with big integers
This adds the algorithm to scalar replace various (for now, only very
carefully constructed) cases of objects boxing big integers. This will
not yet work; the ops it rewrites things into are unimplemented, and the
freeing of any allocated big integers is not yet implemented, so they
will be leaked.