Greatly reduce allocation churn on bigint ops
When we get an operation where we have a big integer and a small big
integer, we have to upgrade the small one to do the operation. We used
to `malloc` and `mp_init` a fresh `mp_int` to use every time, and then
`mp_clear` and `free` it at the end. We can get rid of this cost by
holding a few `mp_int` objects on the `ThreadContext` and re-using them
for all of these operations. This shaves around a third off a Rat
addition benchmark.