Avoid some spurious PEA bailouts
When we reach a merge point in the graph, if the allocation was only
seen on one side and not the other, then it cannot possibly need a
materialization on the side where it wasn't yet created. We don't
insert such materializations yet, but rather bail out; still, fixing
this means we bail out in some cases where we didn't need to do so.