Don't needlessly install VMNulls into ->work
A while back, we put analysis in spesh to make sure that we never read
any object registers before writing them, and in the case that happens
to explicitly `null` them at frame entry, so we don't need to poke null
values in at frame creation time. However, we were still running the
loop to install the VMNulls! Don't do that. Should cheap entry to all
spesh'd/JIT-compiled frames.