[JIT] Template for 'create' should store first then initialize
If we call 'initialize' before storing '$obj' to memory, the '$obj' is
treated as an (anonymous) temporary value, and its location when
spilled to memory is not marked as an object.
If GC is then activated betwen the 'allocate' and the 'initialize'
calls, the (spilled) value of $obj will be pointing to the old
allocated object... and we'll have all sorts of invalid memory bugs.

After we change the order of the store and the call to
initialize (consistent with the order in interp.c), this can never
happen.

Possible future direction - have an 'object' operator that simply
marks its operand as an object.