Make socket family handling portable
MVM_io_resolve_host_name assumes a 64-bit system where the values of the
AF constants are the same as those on Linux. This breaks domain name
resolution on OSes where the values of AF constants aren't the same as
Linux's, such as FreeBSD. In addition, Rakudo never set the family to
the upper bits of the port passed to the function, meaning that its
value could be anything depending on the compiler.

This makes the socket family a separate parameter from the port and uses
platform-independent socket family constants, which are translated to
the correct AF constant by MVM_io_resolve_host_name to use with
getaddrinfo.

Related to https://github.com/rakudo/rakudo/issues/3007