[JIT] Handle NaN in FP branching comparisons
Adds branches checking for parity flag, because the carry flag is also
set when the result of comparison is unordered (i.e. one operand is a
NaN value).

The reason I never saw them in C compiler output is that the C
compiler just reorders the comparison, so it can always use the
is-greater comparison result ('ja' jump). That might be an interesting
optimization for us as well.