Put limits on local and inline counts for inlining
We have some non-linear algorithms as part of inlining - of note, in the
deopt tracking. If the frame gets too big in various ways, the cost of
that can become very large. Put some limits in place to catch such
cases, so we won't end up spending a vast amount of time trying to win
performance, but end up blocking the application for multiple seconds
due to the spesh thread getting stuck in a highly costly loop.