Fix autobox possibly writing to an outdated target
Since autobox is a macro, the target argument is not evaluated at the place of
the call but in the expanded code. That happens only after allocating, so the
GC may have already moved the target frame. Avoid this by forgoing the local
variable and accessing tc->cur_frame->caller again instead as that will have
been updated.