Fix slurpy arg handling possibly accessing an outdated arg context
The arg context is inlined into the frame, so even though its not an
MVMCollectable itself it may move during GC.

MVM_args_slurpy_named has 2 callers. For param_sn we want to access
tc->cur_frame->params directly as that's gonna be up to date. The call capture
in capturenamedshash does not inline the arg context, so we can use what we
get passed without problems.