Fix comp units using obsolete pointers to their filename
CompUnits are allocated directly in gen2 while the filename that ends up in
cu->body.filename usually starts out in the nursery. So in every place we set
that filename property we need a write barrier. Otherwise the reference won't
get updated when the filename will be moved by the GC.