Fix panic in spesh when deserialization causes GC during fact finding
Deserialization may trigger garbage collection or at least allow for it to
happen, but in spesh there are very few places where GC is actually allowed.
If a WVal is not yet deserialized at the time the code its used in is speshed
then we may assume that it's used by a cold branch inside of a hot routine. In
that case its facts are probably not as important, so we may as well get by
without them. Forgoing those facts gets rid of the GC issue.

Many thanks to timotimo++ for the valuable insights and the solution!