New GC debug level
MVM_GC_DEBUG 3 will trigger the GC on every allocation. In addition it
will poison the past fromspace by overwriting it with 0xef (which is
easily recognizable as garbage and cause explosions when dereferenced).
Note that this will also disable destructors as they rely on STables
remaining in past fromspace till the next GC run