Introduce vmevent, "VM Event Subscription"
Lets a ConcBlockingQueue be configured to receive
a set of events from inside the VM.

Currently there's only "gcevent", which gives you
an object in the queue every time the GC runs. It's
an int64 array containing GC sequence number, absolute
start time, start time since vm started, time taken,
minor or major collection, bytes promoted to the old
generation, and thread id of the thread that coordinated
(almost equivalent to the thread that caused GC to
run in the first pace).