Tweak spesh logging's new compunit handling/quotas
The mechanism to put a new spesh log in place when we see a new
compilation unit is helpful for making sure we do decently at OSR for
program mainlines. However, it was not so well engineered; recent Rakudo
changes made it start to miss out on optimizing some programs, but it
also turned out to be to blame for #968. Tune the mechanism, thus
undoing the losses on Rakudo HEAD and fixing #968 along the way.

Also, threaded applications are common, and one log in flight at a time
is not really ideal; tweak the spesh logs "in flight" limit to help with
that.