Fix stack simulation losing track in some cases
When we had specialized code calling unspecialized code, it was possible
for the stack in the specializer's stack simulation to keep on growing.
This made searches on it far more costly, not to mention increasing
memory use (though bounded to the point that the unspecialized code got
to being specialized, so it wasn't unbounded). Make sure that we log
enough to know to take such frames away again. Fixes #1060.