Fix possible memory corruption in MVM_frame_capture_*
In both functions only the code argument was MVMROOTed, but not the local
variable used to avoid the type casts to MVMCode*. This could lead to writes to
uninitialized areas of the nursery which would later trip up some code that
relied on freshly allocated memory to be zeroed out.
Fixed by removing those local variables. MVMROOTing them would incure a
runtime cost that's not worth a few saved characters of code .