heapsnapshot: add per-file/snapshot metadata entries
the per-file metadata contains the time the recording
started, the PID, a sub-version (currently the version
of the format is 3.1, so subversion is 1), and how the
highscore fields (topID and topscore) work.

the per-snapshot metadata contains the time the snapshot
was recorded, the GC sequence number at the time, the
total heap size, the total number of objects, typeobjects,
STables, frames, and references in the heap.

This last data is what the moar-heapanalyzer used to
spend a bunch of time calculating by itself, but the
code here can calculate all values drastically faster.