[Spesh] Make deopt use index rather than target address
We'd use the index to search for the deopt address during compilation,
which makes the deopt index 'special'. And then we'd use the deopt
address to search for the deopt index, in order to find the objects
that we'd need to rematerialize.

Since we maintain the table anyway, it's quite a bit simpler to use
the deopt_idx as the argument, rather than the target address, and
lookup the relevant addresses during deopt runtime. This works
identically for both JIT and interpreted code. (Does require some
tweaking during inlining, though).

Currently causes the following nqp testcases to fail:
- t/qregex/01-qregex.t
- t/moar/52-pluggable-spesh.t