[JIT] Use correct deopt index for guard opcodes
Due to annotation migration (I think) we can have multiple deopt
annotations on the same opcode, even a guard opcode, and they don't
even always have to point to the same deopt target. But we should use
the deopt target that was chosen when the guard was introduced, which
is just the one that's the opcode argument.

Indeed, that's mostly the point of this set of patches.
This fixes the NQP qregex testcase.