[JIT] Deoptimization in the expression JIT
Now that sp_guard* ops are non-special (because they use the deopt
index directly) it's relatively straightforward to implement them as
expression template. So we did.

It turns out that the deoptimization code doesn't unwind the JIT
stack, so we need an explicit jump-to-exit, which is fine.
We can optimize this stuff by further by taking it out of the mainline
code, but that's interesting for another day.