[JIT] Use the new deopt index for synthetic guards
Even though the new guard index points to the same target, it doesn't
typically have the same deopt_source address as the original one, as
that would've been on a prepargs instruction (or similar).

Using the old index would've given us the right target, but anything
referencing the deopt index on the new annotation (in particular, PEA)
would miss the reference. So we should really use the correct one.

This fixes the bug in rakudo t/spes/S32-str/sprintf-b.t